Fix what the live backend actually did, not what the demo pretended
Pointing the app at the real Supabase project instead of the in-memory
demo turned up a set of faults the demo could never show. The schema was
deployed and the RLS suite passed; the bugs were in the space between
them and the client.

Database (new migration, applied to the live project)

- Discovery returned nothing, for everyone, forever. app.discover
  requires snapped_point on both sides and nothing ever wrote it: the
  only caller of set_my_location ran at the court step, before the
  profile row exists, so the insert failed its foreign key and was
  swallowed. A home court is a location, so it now seeds the point,
  snapped to the same coarse grid as a device fix. A real fix still
  wins when the player grants one.
- Whose move it is was a hint, not a rule. Either participant could
  write any state, so the sender could accept their own request. The
  demo enforced this and the live schema did not.
- "Did you hit?" was answerable once. A second tap, or changing your
  mind, raised a duplicate-key error in the user's face.
- expire_requests and enqueue_tomorrow_reminders were callable by any
  signed-in user, because Postgres grants EXECUTE to PUBLIC by default.
  One of them expires every open request in the database. Revoked, and
  default privileges now revoke it so the next one cannot happen.
- A parent could not read the player their child would be meeting, so
  the approval screen rendered "undefined hits played". That is the
  whole safety promise. Scoped policy: the counterparty, on their own
  child's hit, while it is live.
- peek_cohort and roster_counts, because both numbers were being
  counted through a table whose RLS hid the rows.

Client

- Session restore could brick the app: a network blip in the restore
  threw past setReady, the splash was already gone, and the user sat
  looking at an empty green screen with no error and no retry.
- app.json declared neither expo-image-picker nor expo-location, so iOS
  terminated on the first photo (no usage strings) and Android silently
  denied location. Added both, plus the splash plugin, and blocked the
  microphone and storage permissions that came in behind them. An app
  for teenagers should not ask for the microphone.
- The peek screen said "0 players around your level" against a real
  backend, which is the opposite of its job. It now counts for real,
  and says "You're early" when the honest answer is nobody yet.
- Requests from a vanished player were a three-field object cast to
  Player, rendering "undefined hits". Now a whole object that says it
  knows nothing.
- Unhandled rejections on accept, decline, looking-to-hit, availability
  and the three photo actions. Every one was a dead control.
- MatchFound called useAnimatedStyle in a loop, and could not fit a
  360x640 phone: "Share it" and "Done" went off the bottom with no way
  to scroll.
- textShadow* is deprecated on web exactly like shadow* was; same
  helper shape. shadow() also fabricated rgba(NaN,NaN,NaN) for any
  colour that was not six-digit hex.
- useAsync held a stale fn and set state after unmount.
- defaultProposal could send court_id: undefined from the main button.
- The availability step sent nulls straight through on a deep link,
  which classified a minor as an adult.
- The demo backend's load() raced with itself and leaked one permanent
  interval per caller.

UTR

- utr-link wrote the literal string "undefined" into a unique column
  when the profile shape did not match the guesses, so the second
  player to link collided. It now refuses instead, reads field names
  from UTR_PROFILE_FIELDS so the real docs can be applied without a
  redeploy, fixes the hits:/// triple-slash redirect, and the app reads
  the callback outcome instead of treating every failure as success.

Test harness

- scripts/db-test.sh anchored its error check at ^ERROR, but psql
  prefixes errors with the file and line, so a test that died mid-file
  just reported fewer assertions and still said PASSED. It now catches
  those and checks the plan count, which immediately surfaced a real
  pre-existing failure in the UTR test.

125 assertions pass.

+++ supabase/tests/10_utr_link.sql
@@ -11,7 +11,8 @@ select lives_ok($$ select app.begin_utr_link() $$, 'a client can start the link 
 reset role;
 select app.apply_utr(:'adult_a'::uuid, 'utr-123', 9.52, 'rated', 'tok', 'ref');
 select is((select level_source from app.profiles where id = :'adult_a'::uuid), 'utr_verified'::app.level_source, 'a rated player becomes utr_verified');
-select is((select level_value from app.profiles where id = :'adult_a'::uuid), 9.52::numeric, 'and takes the UTR rating');
+-- level_value is a domain over numeric, so the cast is what lets pgTAP resolve is().
+select is((select level_value::numeric from app.profiles where id = :'adult_a'::uuid), 9.52::numeric, 'and takes the UTR rating');
 select ok(not has_table_privilege('authenticated', 'app.utr_links', 'SELECT'), 'tokens are unreadable by clients');
 select * from finish();
 rollback;
